moved localp index manipulatons to a separate file
