+++ SparseGrids/tsgIndexManipulator.cpp
@@ -317,105 +317,6 @@ Data2D<int> computeDAGup(MultiIndexSet const &mset){
     return parents;
 }
 
-Data2D<int> computeDAGup(MultiIndexSet const &mset, const BaseRuleLocalPolynomial *rule){
-    size_t num_dimensions = mset.getNumDimensions();
-    int num_points = mset.getNumIndexes();
-    if (rule->getMaxNumParents() > 1){ // allow for multiple parents and level 0 may have more than one node
-        int max_parents = rule->getMaxNumParents() * (int) num_dimensions;
-        Data2D<int> parents(max_parents, num_points, -1);
-        int level0_offset = rule->getNumPoints(0);
-        #pragma omp parallel for schedule(static)
-        for(int i=0; i<num_points; i++){
-            const int *p = mset.getIndex(i);
-            std::vector<int> dad(num_dimensions);
-            std::copy_n(p, num_dimensions, dad.data());
-            int *pp = parents.getStrip(i);
-            for(size_t j=0; j<num_dimensions; j++){
-                if (dad[j] >= level0_offset){
-                    int current = p[j];
-                    dad[j] = rule->getParent(current);
-                    pp[2*j] = mset.getSlot(dad);
-                    while ((dad[j] >= level0_offset) && (pp[2*j] == -1)){
-                        current = dad[j];
-                        dad[j] = rule->getParent(current);
-                        pp[2*j] = mset.getSlot(dad);
-                    }
-                    dad[j] = rule->getStepParent(current);
-                    if (dad[j] != -1){
-                        pp[2*j + 1] = mset.getSlot(dad);
-                    }
-                    dad[j] = p[j];
-                }
-            }
-        }
-        return parents;
-    }else{ // this assumes that level zero has only one node
-        Data2D<int> parents((int) num_dimensions, num_points);
-        #pragma omp parallel for schedule(static)
-        for(int i=0; i<num_points; i++){
-            const int *p = mset.getIndex(i);
-            std::vector<int> dad(num_dimensions);
-            std::copy_n(p, num_dimensions, dad.data());
-            int *pp = parents.getStrip(i);
-            for(size_t j=0; j<num_dimensions; j++){
-                if (dad[j] == 0){
-                    pp[j] = -1;
-                }else{
-                    dad[j] = rule->getParent(dad[j]);
-                    pp[j] = mset.getSlot(dad.data());
-                    while((dad[j] != 0) && (pp[j] == -1)){
-                        dad[j] = rule->getParent(dad[j]);
-                        pp[j] = mset.getSlot(dad);
-                    }
-                    dad[j] = p[j];
-                }
-            }
-        }
-        return parents;
-    }
-}
-
-Data2D<int> computeDAGDown(MultiIndexSet const &mset, const BaseRuleLocalPolynomial *rule){
-    size_t num_dimensions = mset.getNumDimensions();
-    int max_1d_kids = rule->getMaxNumKids();
-    int num_points = mset.getNumIndexes();
-    Data2D<int> kids(Utils::size_mult(max_1d_kids, num_dimensions), num_points);
-
-    #pragma omp parallel for
-    for(int i=0; i<num_points; i++){
-        std::vector<int> kid(num_dimensions);
-        std::copy_n(mset.getIndex(i), num_dimensions, kid.data());
-        auto family = kids.getIStrip(i);
-
-        for(size_t j=0; j<num_dimensions; j++){
-            int current = kid[j];
-            for(int k=0; k<max_1d_kids; k++){
-                kid[j] = rule->getKid(current, k);
-                *family++ = (kid[j] == -1) ? -1 : mset.getSlot(kid);
-            }
-            kid[j] = current;
-        }
-    }
-
-    return kids;
-}
-
-std::vector<int> computeLevels(MultiIndexSet const &mset, BaseRuleLocalPolynomial const *rule){
-    size_t num_dimensions = mset.getNumDimensions();
-    int num_points = mset.getNumIndexes();
-    std::vector<int> level((size_t) num_points);
-    #pragma omp parallel for schedule(static)
-    for(int i=0; i<num_points; i++){
-        const int *p = mset.getIndex(i);
-        int current_level = rule->getLevel(p[0]);
-        for(size_t j=1; j<num_dimensions; j++){
-            current_level += rule->getLevel(p[j]);
-        }
-        level[i] = current_level;
-    }
-    return level;
-}
-
 MultiIndexSet selectFlaggedChildren(const MultiIndexSet &mset, const std::vector<bool> &flagged, const std::vector<int> &level_limits){
     size_t num_dimensions = mset.getNumDimensions();
